refactor: replace deprecated loader_d component with updated loader
- Updated multiple components to use the new `Loader` component instead of the deprecated `loader_d`, ensuring consistent styling and functionality across the application.
- Adjusted the `Loader` usage to include variant and size props for better visual alignment with design specifications.

These changes enhance the maintainability and visual consistency of loading indicators throughout the dashboard.
